requirements.txt
fastapi==0.115.8
uvicorn[standard]==0.30.6
python-dotenv==1.0.1
email-validator==2.2.0
